requirements.txt
anyio==3.6.2
async-timeout==4.0.2
certifi==2022.12.7
click==8.1.3
Flask==2.2.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
redis==4.5.4
rfc3986==1.5.0
sniffio==1.3.0
Werkzeug==2.2.3
